requirements.txt
fastapi[all]
aiohttp
pytest
pytest-repeat
sqlalchemy
alembic
asyncpg
sqlalchemy-utils
pyjwt